added a new test controller
